Split source files by classes.
