+++ src/Tests/SenseNet.BlobStorage.IntegrationTests/BlobStorageIntegrationTests.cs
@@ -1,13 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Data;
 using System.Data.Common;
 using System.Data.SqlClient;
-using System.Diagnostics.PerformanceData;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using IO = System.IO;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SenseNet.Configuration;
@@ -17,11 +13,8 @@ using SenseNet.ContentRepository.Search;
 using SenseNet.ContentRepository.Security;
 using SenseNet.ContentRepository.Storage;
 using SenseNet.ContentRepository.Storage.Data;
-using SenseNet.ContentRepository.Storage.Data.SqlClient;
 using SenseNet.ContentRepository.Storage.Security;
-using SenseNet.Tests;
 using SenseNet.Tests.Implementations;
-using ContentType = System.Net.Mime.ContentType;
 
 namespace SenseNet.BlobStorage.IntegrationTests
 {
@@ -370,32 +363,5 @@ namespace SenseNet.BlobStorage.IntegrationTests
         }
 
         #endregion
-    }
-    internal static class DbReaderExtensions
-    {
-        public static long GetSafeInt64(this IDataReader reader, int index)
-        {
-            return reader.IsDBNull(index) ? 0 : reader.GetInt64(index);
-        }
-        public static int GetSafeInt32(this IDataReader reader, int index)
-        {
-            return reader.IsDBNull(index) ? 0 : reader.GetInt32(index);
-        }
-        public static DateTime? GetSafeDateTime(this IDataReader reader, int index)
-        {
-            return reader.IsDBNull(index) ? (DateTime?)null : reader.GetDateTime(index);
-        }
-        public static string GetSafeString(this IDataReader reader, int index)
-        {
-            return reader.IsDBNull(index) ? null : reader.GetString(index);
-        }
-        public static bool GetSafeBoolFromBit(this IDataReader reader, int index)
-        {
-            return !reader.IsDBNull(index) && reader.GetBoolean(index);
-        }
-        public static byte[] GetSafeBytes(this IDataReader reader, int index)
-        {
-            return reader.IsDBNull(index) ? null : (byte[])reader[index];
-        }
     }
 }
